refactor: centralize node alias resolution
